requirements.txt
# ===================================
# Visual Study Tool - Requirements
# ===================================

# Core Dependencies
gradio>=4.0.0
pillow>=10.0.0
requests>=2.31.0

# Search Engine
duckduckgo-search>=4.0.0
ddgs>=9.10.0

# AI/ML - Computer Vision
sentence-transformers>=2.2.2
torch>=2.0.0
torchvision>=0.15.0

# Testing
pytest>=7.0
pytest-cov>=4.0
